move stuff around, add a high-level integration test for category2openapi
